requirements.txt
torch~=2.6.0
fastapi~=0.115.12
torchvision~=0.21.0
pillow~=11.2.1
uvicorn~=0.23.2
python-multipart~=0.0.20